(nearly) minimal changes for warning-free compilation on Linux, mostly signed-vs-unsigned, followed by unused variables, followed by class member initialisation order restoring removed (unused) function calls and commenting them out instead as requested

+++ OutputGeometryObject.C
@@ -75,7 +75,7 @@ OutputGeometryObject::compute(
         MDataHandle geosHandle = objectHandle.child(AssetNode::outputGeos);
         MArrayDataHandle geoArrayHandle(geosHandle);
         MArrayDataBuilder geosBuilder = geoArrayHandle.builder();
-        if(geosBuilder.elementCount() != myObjectInfo.geoCount)
+        if(geosBuilder.elementCount() != (unsigned int)(myObjectInfo.geoCount))
         {
             needToSyncOutputs = true;
         }
